move all utils with dependencies to helper file
