+++ jest.config.js
@@ -2,9 +2,9 @@ module.exports = {
   preset: 'ts-jest',
   roots: ['<rootDir>/src'],
   testEnvironment: 'node',
-  setupFilesAfterEnv: ['<rootDir>/src/lib/utils/testSetup.ts'],
+  setupFilesAfterEnv: ['<rootDir>/src/lib/__tests__/setup.ts'],
   testPathIgnorePatterns: ['/node_modules/'],
-  testMatch: ['**/__tests__/**/*.+(ts|tsx|js)', '**/?(*.)+(spec|test).+(ts|tsx|js)'],
+  testMatch: ['**/?(*.)+(spec|test).+(ts|tsx|js)'],
   transform: {
     '^.+\\.(ts|tsx)?$': 'ts-jest',
   },
