+++ src/lib/helpers.ts
@@ -1,7 +1,8 @@
+import { Transform } from 'class-transformer';
 import { validateOrReject } from 'class-validator';
 
-import { ConnectionInvitationMessage } from '../protocols/connections/ConnectionInvitationMessage';
-import { MessageTransformer } from '../agent/MessageTransformer';
+import { ConnectionInvitationMessage } from './protocols/connections/ConnectionInvitationMessage';
+import { MessageTransformer } from './agent/MessageTransformer';
 
 /**
  * Create a `ConnectionInvitationMessage` instance from the `c_i` parameter of an URL
@@ -39,4 +40,37 @@ export function encodeInvitationToUrl(
   const invitationUrl = `${domain}?c_i=${encodedInvitation}`;
 
   return invitationUrl;
+}
+
+/**
+ * Provide a default value for a parameter when using class-transformer
+ *
+ * Class transfomer doesn't use the default value of a property when transforming an
+ * object using `plainToClass`. This decorator allows to set a default value when no value is
+ * present during transformation.
+ *
+ * @param defaultValue the default value to use when there is no value present during transformation
+ * @see https://github.com/typestack/class-transformer/issues/129#issuecomment-425843700
+ *
+ * @example
+ * import { plainToClass } from 'class-transformer'
+ *
+ * class Test {
+ *  // doesn't work
+ *  myProp = true;
+ *
+ *  // does work
+ *  ＠Default(true)
+ *  myDefaultProp: boolean;
+ * }
+ *
+ * plainToClass(Test, {})
+ * // results in
+ * {
+ *   "myProp": undefined,
+ *   "myDefaultProp": true
+ * }
+ */
+export function Default<T>(defaultValue: T) {
+  return Transform((value: T | null | undefined) => (value !== null && value !== undefined ? value : defaultValue));
 }
