+++ src/lib/index.ts
@@ -4,7 +4,7 @@ import 'reflect-metadata';
 export { Agent } from './agent/Agent';
 export { InboundTransporter } from './transport/InboundTransporter';
 export { OutboundTransporter } from './transport/OutboundTransporter';
-export { encodeInvitationToUrl, decodeInvitationFromUrl } from './utils/invitationUrl';
+export { encodeInvitationToUrl, decodeInvitationFromUrl } from './helpers';
 
 export { ConnectionRecord } from './storage/ConnectionRecord';
 export { EventType as ConnectionEventType } from './protocols/connections/ConnectionService';
